+++ h2o-algos/src/jmh/java/hex/pca/PCAImputeMissingTrainingBench.java
@@ -1,7 +1,6 @@
 package hex.pca;
 
 import hex.DataInfo;
-import hex.svd.SVDImplementation;
 import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.RunnerException;
