+++ h2o-algos/src/jmh/java/hex/pca/PCAJMHScoring.java
@@ -1,6 +1,5 @@
 package hex.pca;
 
-import hex.svd.SVDImplementation;
 import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.RunnerException;
