+++ h2o-algos/src/jmh/java/hex/pca/PCAWideDataSets.java
@@ -1,7 +1,6 @@
 package hex.pca;
 
 import hex.DataInfo;
-import hex.svd.SVDImplementation;
 import water.DKV;
 import water.Key;
 import water.fvec.Frame;
